requirements.txt
pybullet>=3.2.0
PyYAML>=6.0
numpy
imageio>=2.31
